+++ web-frontend/modules/database/components/row/RowEditFieldSingleSelectReadOnly.vue
@@ -1,12 +1,15 @@
 <template>
-  <RowEditFieldSingleSelect :read-only="true" :field="field">
+  <RowEditFieldSingleSelect :read-only="true" :field="field" :value="value">
   </RowEditFieldSingleSelect>
 </template>
 
 <script>
 import RowEditFieldSingleSelect from '@baserow/modules/database/components/row/RowEditFieldSingleSelect'
+import baseField from '@baserow/modules/database/mixins/baseField'
+
 export default {
   name: 'RowEditFieldSingleSelectReadOnly',
   components: { RowEditFieldSingleSelect },
+  mixins: [baseField],
 }
 </script>